refactor: split system-diagnostics page into focused modules
1361-line monolith split into 9 files (avg ~147 lines each):
- diagnostics-types.ts — 9 API response interfaces
- diagnostics-explanations.ts — 6 explanation dictionaries
- diagnostics-utils.ts — formatBytes (null-safe), formatUptime, formatMicroseconds, formatCount, downloadFromResponse
- stat-grid.tsx — Stat component + StatGrid wrapper (owns responsive-column logic)
- use-diagnostics.ts — 3 useQuery hooks + peek/inspect state
- diagnostics-sections.tsx — ProcessSection, LinuxProcessSection, SmapsRollupSection, HeapSection, ResourceUsageSection
- diagnostics-tables.tsx — HeapSpacesTable, SmapsTopTable
- region-inspect.tsx — RegionInspectPanel
- page.tsx — thin orchestrator (~80 lines of layout + download handlers)

Primary duplication eliminated: `description={showExplanations ? DICT.Key : undefined}`
repeated 50 times → replaced by a per-section `explain(key)` helper.
